requirements.txt
pytest>=7.0
pyserial>=3.0
requests>=2.28
